Added: Edit and delete end points all tested and working, do kindly do some testing on your own systems as well.
Edited:
Removed unused imports and changes to endpoints and params for edit and delete endpoints.

+++ src/main/java/com/fdm/pod/pod/service/PostService.java
@@ -1,6 +1,7 @@
 package com.fdm.pod.pod.service;
 
 import java.util.List;
+import java.util.Optional;
 
 import com.fdm.pod.pod.model.Post;
 import com.fdm.pod.pod.exception.PostDoesNotExistException;
@@ -36,7 +37,12 @@ public class PostService {
         }
     }
 
-    public void deletePost(Long postId){
-        postRepository.deleteById(postId);
+    public boolean deletePost(Long postId){
+        Optional<Post> post = postRepository.findById(postId);
+        if (post.isPresent()){
+            postRepository.deleteById(postId);
+            return true;
+        }
+        return false;
     }
 }
